+++ platform/src/codatplatform/models/shared/branding.py
@@ -13,7 +13,6 @@ from typing import Optional
 
 @dataclasses.dataclass
 class Branding:
-    r"""OK"""
     button: Optional[shared_brandingbutton.BrandingButton] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('button'), 'exclude': lambda f: f is None }})
     r"""Button branding references."""
     logo: Optional[shared_brandinglogo.BrandingLogo] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('logo'), 'exclude': lambda f: f is None }})